requirements.txt
python-dotenv==1.0.1
gql[all]==3.5.0
graphql-core==3.2.3
